requirements_base.txt
otree>=X.X.X
django-recaptcha
